# Optimizations for Ascii.Equals and Ascii.EqualsIgnoreCase