Adresses the open points / comments from https://github.com/dotnet/runtime/pull/84886 (cf. https://github.com/dotnet/runtime/pull/84886#issuecomment-1518242089)

* replaced usage of `Vector64` hw-accelerated intrinsics (instead of sw-fallback), cf. https://github.com/dotnet/runtime/pull/84886#discussion_r1168627003, https://github.com/dotnet/runtime/pull/84886#discussion_r1168629066
* removed bound checks in scalar paths, cf. https://github.com/dotnet/runtime/pull/84886#discussion_r1168566536
* vectorized EqualsIgnoreCase, cf. https://github.com/dotnet/runtime/pull/84886#discussion_r1168586648
* deduplicated the `(byte, char)` and `(byte, byte)` / `(char, char)` variants