@gfoidl 2023-05-10
> Could you please run the benchmarks from [dotnet/performance#3016](https://github.com/dotnet/performance/pull/3016) and share the results?

Sure. In [runtime workflow](https://github.com/dotnet/performance/blob/main/docs/benchmarking-workflow-dotnet-runtime.md) I don't see some instructions on how to compare that directly. In some PRs here I see benchmarks that compare _main_ with _pr_, please give me some advice on how to do these.
Right now I have two builds of runtime (main and for the PR-branch).