@gfoidl 2023-05-11
> Could you please run the benchmarks from [dotnet/performance#3016](https://github.com/dotnet/performance/pull/3016) and share the results?

<details>
  <summary>Machine info</summary>

``` ini

BenchmarkDotNet=v0.13.2.2052-nightly, OS=Windows 10 (10.0.19045.2846)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET SDK=8.0.100-preview.5.23260.3
  [Host]     : .NET 8.0.0 (8.0.23.25213), X64 RyuJIT AVX2
  main : .NET 8.0.0 (42.42.42.42424), X64 RyuJIT AVX2
  pr : .NET 8.0.0 (42.42.42.42424), X64 RyuJIT AVX2

PowerPlanMode=00000000-0000-0000-0000-000000000000  Arguments=/p:EnableUnsafeBinaryFormatterSerialization=true  IterationTime=250.0000 ms
MaxIterationCount=20  MinIterationCount=15  WarmupCount=1
```
</details>

| Method                                      | Job  | Size | Mean       | Error     | Ratio |
|---------------------------------------------|------|------|------------|-----------|-------|
| IsValid_Bytes                               | main | 6    | 2.537 ns   | 0.2359 ns | 1.00  |
| IsValid_Bytes                               | pr   | 6    | 2.016 ns   | 0.0397 ns | 0.80  |
|                                             |      |      |            |           |       |
| IsValid_Chars                               | main | 6    | 3.387 ns   | 0.2208 ns | 1.00  |
| IsValid_Chars                               | pr   | 6    | 2.742 ns   | 0.0860 ns | 0.82  |
|                                             |      |      |            |           |       |
| Equals_Bytes                                | main | 6    | 8.860 ns   | 0.4392 ns | 1.00  |
| Equals_Bytes                                | pr   | 6    | 5.099 ns   | 0.1212 ns | 0.58  |
|                                             |      |      |            |           |       |
| Equals_Chars                                | main | 6    | 9.591 ns   | 0.3690 ns | 1.00  |
| Equals_Chars                                | pr   | 6    | 5.742 ns   | 0.0648 ns | 0.60  |
|                                             |      |      |            |           |       |
| Equals_Bytes_Chars                          | main | 6    | 9.366 ns   | 0.5560 ns | 1.00  |
| Equals_Bytes_Chars                          | pr   | 6    | 6.061 ns   | 0.1365 ns | 0.66  |
|                                             |      |      |            |           |       |
| EqualsIgnoreCase_ExactlyTheSame_Bytes       | main | 6    | 9.026 ns   | 0.4026 ns | 1.00  |
| EqualsIgnoreCase_ExactlyTheSame_Bytes       | pr   | 6    | 8.822 ns   | 0.1208 ns | 0.97  |
|                                             |      |      |            |           |       |
| EqualsIgnoreCase_ExactlyTheSame_Chars       | main | 6    | 8.797 ns   | 0.1992 ns | 1.00  |
| EqualsIgnoreCase_ExactlyTheSame_Chars       | pr   | 6    | 9.300 ns   | 0.1225 ns | 1.06  |
|                                             |      |      |            |           |       |
| EqualsIgnoreCase_ExactlyTheSame_Bytes_Chars | main | 6    | 7.201 ns   | 0.1652 ns | 1.00  |
| EqualsIgnoreCase_ExactlyTheSame_Bytes_Chars | pr   | 6    | 9.225 ns   | 0.2208 ns | 1.28  |
|                                             |      |      |            |           |       |
| EqualsIgnoreCase_DifferentCase_Bytes        | main | 6    | 11.150 ns  | 0.2578 ns | 1.00  |
| EqualsIgnoreCase_DifferentCase_Bytes        | pr   | 6    | 10.116 ns  | 0.2073 ns | 0.91  |
|                                             |      |      |            |           |       |
| EqualsIgnoreCase_DifferentCase_Chars        | main | 6    | 11.533 ns  | 0.1493 ns | 1.00  |
| EqualsIgnoreCase_DifferentCase_Chars        | pr   | 6    | 10.151 ns  | 0.1410 ns | 0.88  |
|                                             |      |      |            |           |       |
| EqualsIgnoreCase_DifferentCase_Bytes_Chars  | main | 6    | 11.220 ns  | 0.1978 ns | 1.00  |
| EqualsIgnoreCase_DifferentCase_Bytes_Chars  | pr   | 6    | 10.293 ns  | 0.1117 ns | 0.92  |
|                                             |      |      |            |           |       |
| IsValid_Bytes                               | main | 128  | 2.958 ns   | 0.0841 ns | 1.00  |
| IsValid_Bytes                               | pr   | 128  | 2.956 ns   | 0.0446 ns | 1.00  |
|                                             |      |      |            |           |       |
| IsValid_Chars                               | main | 128  | 4.798 ns   | 0.1310 ns | 1.00  |
| IsValid_Chars                               | pr   | 128  | 5.039 ns   | 0.0931 ns | 1.05  |
|                                             |      |      |            |           |       |
| Equals_Bytes                                | main | 128  | 11.362 ns  | 0.1925 ns | 1.00  |
| Equals_Bytes                                | pr   | 128  | 6.118 ns   | 0.1530 ns | 0.54  |
|                                             |      |      |            |           |       |
| Equals_Chars                                | main | 128  | 15.352 ns  | 0.3357 ns | 1.00  |
| Equals_Chars                                | pr   | 128  | 12.580 ns  | 0.2221 ns | 0.82  |
|                                             |      |      |            |           |       |
| Equals_Bytes_Chars                          | main | 128  | 17.121 ns  | 0.3781 ns | 1.00  |
| Equals_Bytes_Chars                          | pr   | 128  | 13.550 ns  | 0.3093 ns | 0.80  |
|                                             |      |      |            |           |       |
| EqualsIgnoreCase_ExactlyTheSame_Bytes       | main | 128  | 118.966 ns | 1.8113 ns | 1.00  |
| EqualsIgnoreCase_ExactlyTheSame_Bytes       | pr   | 128  | 10.107 ns  | 0.2089 ns | 0.08  |
|                                             |      |      |            |           |       |
| EqualsIgnoreCase_ExactlyTheSame_Chars       | main | 128  | 153.375 ns | 2.8708 ns | 1.00  |
| EqualsIgnoreCase_ExactlyTheSame_Chars       | pr   | 128  | 15.381 ns  | 0.3254 ns | 0.10  |
|                                             |      |      |            |           |       |
| EqualsIgnoreCase_ExactlyTheSame_Bytes_Chars | main | 128  | 109.199 ns | 1.7805 ns | 1.00  |
| EqualsIgnoreCase_ExactlyTheSame_Bytes_Chars | pr   | 128  | 16.945 ns  | 0.4240 ns | 0.16  |
|                                             |      |      |            |           |       |
| EqualsIgnoreCase_DifferentCase_Bytes        | main | 128  | 159.318 ns | 2.9244 ns | 1.00  |
| EqualsIgnoreCase_DifferentCase_Bytes        | pr   | 128  | 16.653 ns  | 0.3459 ns | 0.10  |
|                                             |      |      |            |           |       |
| EqualsIgnoreCase_DifferentCase_Chars        | main | 128  | 158.219 ns | 2.2267 ns | 1.00  |
| EqualsIgnoreCase_DifferentCase_Chars        | pr   | 128  | 28.308 ns  | 0.5351 ns | 0.18  |
|                                             |      |      |            |           |       |
| EqualsIgnoreCase_DifferentCase_Bytes_Chars  | main | 128  | 231.040 ns | 2.1013 ns | 1.00  |
| EqualsIgnoreCase_DifferentCase_Bytes_Chars  | pr   | 128  | 31.340 ns  | 0.5958 ns | 0.14  |

(removed unrelated rows from the table and renamed the job for easier interpretation)